backend/requirements.txt
webrtcvad>=2.0.10
sounddevice>=0.4.6
numpy>=1.24.0
httpx>=0.27.0
openai>=1.30.0
python-dotenv>=1.0.0
